Ref: moving AbstractLogger type into src/logger.ts file.
